Created inputsdValidations utils Refacto input s validation logic. Fix input s initial value bug when the input is required
