+++ packages/strapi-helper-plugin/lib/src/components/InputsIndex/index.js
@@ -38,7 +38,6 @@ const inputs = {
 };
 
 function InputsIndex(props) {
-  const inputValue = props.type === 'checkbox' || props.type === 'toggle' ? props.value || false : props.value || '';
   const type = props.type && !isEmpty(props.addon) ? 'addon' : props.type;
   const inputValue = props.type === 'checkbox' || props.type === 'toggle' ? props.value || false : props.value || '';
   const Input = inputs[type] ? inputs[type] : DefaultInputError;
